ai-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google Gemini
google-generativeai>=0.8.0

# LLM Tooling
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
instructor>=1.7.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0
